+++ rest-demo-jpa-mysql/src/main/java/com/frankaboagye/restdemojpamysql/repositories/CloudVendorRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface CloudVendorRepository extends JpaRepository<CloudVendor, Integer> {
+public interface CloudVendorRepository extends JpaRepository<CloudVendor, Long> {
     List<CloudVendor> findByVendorName(String vendorName);
 }
