+++ rest-demo-jpa-mysql/src/main/java/com/frankaboagye/restdemojpamysql/services/CloudVendorService.java
@@ -7,8 +7,8 @@ import java.util.List;
 public interface CloudVendorService {
 
     public String createCloudVendor(CloudVendor cloudVendor);
-    public String updateCloudVendor(CloudVendor cloudVendor, Integer vendorId);
-    public String deleteCloudVendor(Integer vendorId);
-    public CloudVendor getCloudVendor(Integer vendorId);
+    public String updateCloudVendor(CloudVendor cloudVendor, Long vendorId);
+    public String deleteCloudVendor(Long vendorId);
+    public CloudVendor getCloudVendor(Long vendorId);
     public List<CloudVendor> getAllCloudVendors();
 }
